[IMP] Readability and maintainability
- make one objet in one file
- improve docs
